Revert "Move office hours CTA from footer to front page only"
This reverts commit 82b788550c9ba6eea61843bc7610f4d9cedec883.
