+++ src/layout/footer/index.js
@@ -2,6 +2,7 @@ import { graphql, StaticQuery, Link } from "gatsby"
 import React from "react"
 import MarketoForm from "../../components/marketoForm"
 import {
+  CTASlice,
   InputText,
   FlexContainer,
   Form,
@@ -50,73 +51,92 @@ const Footer = ({ data }) => {
     (file) => file.node.publicURL.indexOf("CC-BY-SA_icon") > -1
   )
 
-  return (
-    <SiteFooter containerWidth="wide">
-      <div className="pds-footer__links-area pds-grid pds-grid--wide">
-        <div className="pds-footer__links-area-column pds-grid-item pds-grid-item--sm-4 pds-grid-item--md-12 pds-grid-item--lg-8">
-          <FlexContainer
-            spacing="standard"
-            justifyContent="between"
-            mobileFlex="reverse"
-          >
-            <FooterLinks
-              headingText="Topics"
-              headingLevel="h2"
-              linkItems={topicLinks}
-            />
-            <FooterLinks
-              headingText="Help"
-              headingLevel="h2"
-              linkItems={helpLinks}
-              className="pds-spacing-mar-block-start-xl@sm"
-            />
-            <FooterLinks
-              headingText="Community"
-              headingLevel="h2"
-              linkItems={communityLinks}
-              className="pds-spacing-mar-block-start-xl@sm"
-            />
-          </FlexContainer>
-        </div>
-        <div className="pds-footer__links-area-column pds-grid-item pds-grid-item--sm-4 pds-grid-item--md-12 pds-grid-item--lg-4">
-          <FooterHeading
-            headingText="Connect"
-            headingLevel="h2"
-            className="pds-spacing-pad-block-start-xl pds-spacing-pad-block-start-none@lg"
-          />
-          <p className="footer-form-intro pds-ts-s">
-            Awesome development news, tutorials, and tips. Plus get 3 free
-            downloads, just for signing up. If you don't love it, unsubscribe
-            with just a click.
-          </p>
-          <div className="pds-spacing-mar-block-xl">
-            <MarketoForm
-              baseUrl="https://app-ab05.marketo.com"
-              munchkinId="316-GSV-089"
-              formId={2014}
-              formName="mktoForm_2014"
-            />
-          </div>
-          <SocialLinks className="pds-spacing-mar-block-start-xl" />
-        </div>
-      </div>
+  const primaryCTA = {
+    text: "Learn Pantheon",
+    url: "https://pantheon.io/learn-pantheon?docs",
+  }
 
-      <div className="cc-license pds-spacing-mar-block-start-5xl">
-        <div className="cc-license__logo">
-          {CCLogo && (
-            <img
-              src={CCLogo.node.publicURL}
-              alt="Creative Commons Attribution-ShareAlike Logo"
+  const secondaryCTA = {
+    text: "Office Hours",
+    url: "https://pantheon.io/developers/office-hours?docs",
+  }
+
+  return (
+    <>
+      <CTASlice
+        backgroundColor="secondary"
+        containerWidth="wide"
+        headingText="Got questions? We've got answers!"
+        primaryLink={primaryCTA}
+        secondaryLink={secondaryCTA}
+      />
+      <SiteFooter containerWidth="wide">
+        <div className="pds-footer__links-area pds-grid pds-grid--wide">
+          <div className="pds-footer__links-area-column pds-grid-item pds-grid-item--sm-4 pds-grid-item--md-12 pds-grid-item--lg-8">
+            <FlexContainer
+              spacing="standard"
+              justifyContent="between"
+              mobileFlex="reverse"
+            >
+              <FooterLinks
+                headingText="Topics"
+                headingLevel="h2"
+                linkItems={topicLinks}
+              />
+              <FooterLinks
+                headingText="Help"
+                headingLevel="h2"
+                linkItems={helpLinks}
+                className="pds-spacing-mar-block-start-xl@sm"
+              />
+              <FooterLinks
+                headingText="Community"
+                headingLevel="h2"
+                linkItems={communityLinks}
+                className="pds-spacing-mar-block-start-xl@sm"
+              />
+            </FlexContainer>
+          </div>
+          <div className="pds-footer__links-area-column pds-grid-item pds-grid-item--sm-4 pds-grid-item--md-12 pds-grid-item--lg-4">
+            <FooterHeading
+              headingText="Connect"
+              headingLevel="h2"
+              className="pds-spacing-pad-block-start-xl pds-spacing-pad-block-start-none@lg"
             />
-          )}
+            <p className="footer-form-intro pds-ts-s">
+              Awesome development news, tutorials, and tips. Plus get 3 free
+              downloads, just for signing up. If you don't love it, unsubscribe
+              with just a click.
+            </p>
+            <div className="pds-spacing-mar-block-xl">
+              <MarketoForm
+                baseUrl="https://app-ab05.marketo.com"
+                munchkinId="316-GSV-089"
+                formId={2014}
+                formName="mktoForm_2014"
+              />
+            </div>
+            <SocialLinks className="pds-spacing-mar-block-start-xl" />
+          </div>
         </div>
-        <p className="cc-license__text pds-ts-s">
-          Our Documentation is licensed under a Creative Commons
-          Attribution-ShareAlike 4.0 International License. Code snippets are
-          additionally licensed under The MIT License.
-        </p>
-      </div>
-    </SiteFooter>
+
+        <div className="cc-license pds-spacing-mar-block-start-5xl">
+          <div className="cc-license__logo">
+            {CCLogo && (
+              <img
+                src={CCLogo.node.publicURL}
+                alt="Creative Commons Attribution-ShareAlike Logo"
+              />
+            )}
+          </div>
+          <p className="cc-license__text pds-ts-s">
+            Our Documentation is licensed under a Creative Commons
+            Attribution-ShareAlike 4.0 International License. Code snippets are
+            additionally licensed under The MIT License.
+          </p>
+        </div>
+      </SiteFooter>
+    </>
   )
 }
